+++ crates/aptos-dkg/tests/range_proof.rs
@@ -21,7 +21,7 @@ use std::any::type_name;
 fn assert_range_proof_correctness<E: Pairing, B: BatchedRangeProof<E>>(
     setup: &RangeProofUniversalSetup<E, B>,
     n: usize,
-    ell: u8,
+    ell: usize,
 ) {
     let mut rng = rand::thread_rng();
     let RangeProofUniversalSetup { pk, vk } = setup;
@@ -116,7 +116,7 @@ fn assert_keys_serialization<E: Pairing, B: BatchedRangeProof<E>>(
 }
 
 #[cfg(test)]
-const TEST_CASES: &[(usize, u8)] = &[
+const TEST_CASES: &[(usize, usize)] = &[
     // (n, \ell)
     (1, 16),
     (3, 16),
@@ -139,7 +139,7 @@ struct RangeProofUniversalSetup<E: Pairing, B: BatchedRangeProof<E>> {
 
 #[cfg(test)]
 /// Generate a fixed setup for a single curve
-fn make_single_curve_setup<E, B>(n: usize, ell: u8) -> RangeProofUniversalSetup<E, B>
+fn make_single_curve_setup<E, B>(n: usize, ell: usize) -> RangeProofUniversalSetup<E, B>
 where
     E: Pairing,
     B: BatchedRangeProof<E>,
